Update and rename scalex14.md to volunteer.md
I decided to change the website tree so there is only one volunteer tab and it will be the one for people to fill out the volunteer form. I am going to post the scale volunteer information as a blog so people see it as soon as they travel to the site.